sinwebapp/requirements.txt
django>=2.2.13
cg-django-uaa==2.0.0
psycopg2==2.8.5
gunicorn==20.0.4
django-cors-headers==3.4.0
pyjwt==1.7.1